requirements.txt
attrs==21.2.0
importlib-metadata==4.6.4
imutils==0.5.4
iniconfig==1.1.1
numpy==1.19.5
opencv-contrib-python==4.5.3.56
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-timeout==1.4.2
pyzmp==0.0.17
pyzmq==22.2.1
six==1.16.0
tblib==1.7.0
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.5.0
